ssh_key_loglevel_test.go
package ssh_config

import "testing"

// LogLevel
//         Gives the verbosity level that is used when logging messages
//         from ssh(1).  The possible values are: QUIET, FATAL, ERROR,
//         INFO, VERBOSE, DEBUG, DEBUG1, DEBUG2, and DEBUG3.  The default
//         is INFO.  DEBUG and DEBUG1 are equivalent.  DEBUG2 and DEBUG3
//         each specify higher levels of verbose output.
//

func TestLogLevelInitFunc(t *testing.T) {
	defer func() {
		if err := recover(); err != nil {
			t.Errorf("panic in test is not acceptable")
		}
	}()
	f, ok := mapInit[LogLevel]
	if !ok {
		t.Errorf("Cannot found init-function")
	}
	if f == nil {
		t.Errorf("Init-function is nil")
	}
	if f() == "" {
		// TODO
		t.Errorf("Not acceptable empty init value")
	}
}

func TestLogLevelValidFunc(t *testing.T) {
	defer func() {
		if err := recover(); err != nil {
			t.Errorf("panic in test is not acceptable")
		}
	}()
	f, ok := mapValid[LogLevel]
	if !ok {
		t.Errorf("Cannot found valid-function")
	}
	if f == nil {
		t.Errorf("Valid-function is nil")
	}
	if !f("") {
		// TODO
		t.Errorf("Not acceptable empty valid-value")
	}
}

func TestLogLevelParseFunc(t *testing.T) {
	defer func() {
		if err := recover(); err != nil {
			t.Errorf("panic in test is not acceptable")
		}
	}()
	f, ok := mapParse[LogLevel]
	if !ok {
		t.Errorf("Cannot found parse-function")
	}
	if f == nil {
		t.Errorf("Parse-function is nil")
	}
	// TODO
}
